New components
* Added Carousel file structure;
* Carousel card optimized for responsive display;
* Few animations added;
