--- components/home/newsCarousel.js
@@ -1,9 +1,0 @@
-import { ViewportHook } from "../viewportHook";
-import { NewsCardCarousel } from "./newsCard";
-import React from "react";
-
-export default function NewsCarousel() {
-  const isBreakpoint = ViewportHook(850); //mobile
-
-  return <div></div>;
-}
