+++ package.json
@@ -30,6 +30,7 @@
     "react-focus-lock": "^2.9.2",
     "react-icons": "^4.6.0",
     "react-modern-drawer": "^1.1.2",
+    "react-responsive-carousel": "^3.2.23",
     "styled-components": "^5.3.6",
     "typescript": "4.9.3"
   },
